add storybook and restore file
